fix: change reactivity handling in db
There are a few parts in which it is necessary to handle -999 values as
NAN and vice versa:
- db contains -999
- MASS algorithm requires -999 instead of NANs

Reactivities from DB are now a type that wraps a reactivity types and
treats placeholder values (-999) like NANs. In this way it is possible
to avoid allocations or changes to the underlying data.